backend/pyproject.toml
[project]
name = "anak-tournaments-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiobotocore>=2.17.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.1",
    "cashews>=7.4.0",
    "celery>=5.4.0",
    "fastapi[standard]>=0.115.6",
    "flower>=2.0.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "orjson>=3.10.14",
    "passlib>=1.7.4",
    "psycopg[binary]>=3.2.3",
    "pydantic-settings>=2.7.1",
    "pydantic[email]>=2.10.5",
    "pyjwt>=2.10.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "regex>=2024.11.6",
    "sentry-sdk>=2.20.0",
    "setuptools>=75.8.0",
    "sqlalchemy>=2.0.37",
    "sqlalchemy-utils>=0.41.2",
    "uvicorn>=0.34.0",
    "uvloop; platform_system == 'Linux'",
]


[tool.uv]
dev-dependencies = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py312"
exclude = ["alembic", "app/schemas/__init__.py", "app/models/__init__.py"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
#    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
    "F403", # 'from module import *' used; unable to detect undefined names
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
